@@ -16,34 +16,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.7",
-    "@angular/cdk": "^6.3.3",
-    "@angular/common": "^6.0.7",
-    "@angular/compiler": "^6.0.7",
-    "@angular/core": "^6.0.7",
-    "@angular/forms": "^6.0.7",
-    "@angular/http": "^6.0.7",
-    "@angular/material": "^6.3.3",
-    "@angular/material-moment-adapter": "^6.4.7",
-    "@angular/platform-browser": "^6.0.7",
-    "@angular/platform-browser-dynamic": "^6.0.7",
-    "@angular/router": "^6.0.7",
-    "fedeeman-datetimepicker": "git+https://github.com/fedeeman/fedeeman-datetimepicker.git#1.0.0",
-    "core-js": "^2.5.7",
-    "moment": "^2.18.1",
-    "rxjs": "^6.2.1",
-    "zone.js": "^0.8.26"
+    "@angular/animations": "^9.0.6",
+    "@angular/cdk": "^9.1.2",
+    "@angular/common": "^9.0.6",
+    "@angular/compiler": "^9.0.6",
+    "@angular/core": "9.0.6",
+    "@angular/forms": "^9.0.6",
+    "@angular/material": "^9.1.2",
+    "@angular/material-moment-adapter": "^9.1.2",
+    "@angular/platform-browser": "^9.0.6",
+    "@angular/platform-browser-dynamic": "^9.0.6",
+    "@angular/router": "^9.0.6",
+    "core-js": "^3.6.4",
+    "moment": "^2.24.0",
+    "rxjs": "^6.5.4",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.8.4",
-    "@angular-devkit/build-ng-packagr": "~0.8.4",
-    "@angular/cli": "~6.2.4",
-    "@angular/compiler-cli": "^6.1.9",
-    "@angular/language-service": "^6.0.1",
+    "@angular-devkit/build-angular": "~0.900.6",
+    "@angular-devkit/build-ng-packagr": "~0.900.6",
+    "@angular/cli": "~9.0.6",
+    "@angular/compiler-cli": "^9.0.6",
+    "@angular/language-service": "^9.0.6",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "copy-files-from-to": "0.2.3",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
@@ -52,13 +51,11 @@
     "karma-coverage-istanbul-reporter": "~2.0.0",
     "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "3.0.6",
+    "ng-packagr": "^9.0.0",
     "protractor": "~5.3.0",
     "rimraf": "2.6.2",
     "ts-node": "~5.0.1",
-    "tsickle": ">=0.25.5",
-    "tslib": "^1.7.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.9.1"
+    "typescript": "~3.7.5"
   }
 }
