@@ -1,12 +1,12 @@
 {
   "name": "fedeeman-datetimepicker/core",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "repository": "https://github.com/fedeeman/fedeeman-datetimepicker.git",
   "author": "JF <j.fedor38@gmail.com>",
   "license": "MIT",
   "peerDependencies": {
-    "@angular/core": "^6.0.7",
-    "@angular/material": "^6.3.3",
-    "@angular/cdk": "^6.3.3"
+    "@angular/core": "^9.0.0",
+    "@angular/material": "^9.0.0",
+    "@angular/cdk": "^9.0.0"
   }
 }
