@@ -1,8 +1,8 @@
 import { NgModule } from "@angular/core";
 import {
   MatNativeDateModule,
-  NativeDateModule
-} from "@angular/material";
+  NativeDateModule,
+} from "@angular/material/core";
 import { DatetimeAdapter } from "./datetime-adapter";
 import { MAT_DATETIME_FORMATS } from "./datetime-formats";
 import { NativeDatetimeAdapter } from "./native-datetime-adapter";
