@@ -1,17 +1,2 @@
-export * from "./adapter";
-export * from "./adapter/adapter.module";
-export * from "./adapter/datetime-adapter";
-export * from "./adapter/datetime-formats";
-export * from "./adapter/native-datetime-adapter";
-export * from "./adapter/native-datetime-formats";
-
-export * from "./datetimepicker";
-export * from "./datetimepicker/datetimepicker.module";
-export * from "./datetimepicker/calendar";
-export * from "./datetimepicker/calendar-body";
-export * from "./datetimepicker/datetimepicker";
-export * from "./datetimepicker/datetimepicker-filtertype";
-export * from "./datetimepicker/datetimepicker-input";
-export * from "./datetimepicker/datetimepicker-toggle";
-export * from "./datetimepicker/month-view";
-export * from "./datetimepicker/year-view";
+export * from "./adapter/index";
+export * from "./datetimepicker/index";
