@@ -4,14 +4,14 @@ import {
   FormGroup,
   Validators
 } from "@angular/forms";
-import { DateAdapter } from "@angular/material";
+import { DateAdapter } from "@angular/material/core";
 import { MomentDateAdapter } from "@angular/material-moment-adapter";
 import {
   DatetimeAdapter,
   MAT_DATETIME_FORMATS,
   MatDatetimepickerFilterType
-} from "@mat-datetimepicker/core";
-import { MomentDatetimeAdapter } from "@mat-datetimepicker/moment";
+} from "fedeeman-datetimepicker/core";
+import { MomentDatetimeAdapter } from "fedeeman-datetimepicker/moment";
 import {
   Moment,
   utc
