@@ -1,13 +1,11 @@
 import { NgModule } from "@angular/core";
 import { ReactiveFormsModule } from "@angular/forms";
-import {
-  MatDatepickerModule,
-  MatFormFieldModule,
-  MatInputModule
-} from "@angular/material";
+import { MatDatepickerModule } from "@angular/material/datepicker";
+import { MatFormFieldModule } from "@angular/material/form-field";
+import { MatInputModule } from "@angular/material/input";
 import { BrowserModule } from "@angular/platform-browser";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
-import { MatDatetimepickerModule } from "@mat-datetimepicker/core";
+import { MatDatetimepickerModule } from "fedeeman-datetimepicker/core";
 import { NativeDatetimeComponent } from "./native.component";
 
 @NgModule({
